+++ src/usedBits.cpp
@@ -108,7 +108,7 @@ void ENode::passWidthToChild() {
       childBits.push_back(usedBit);
       break;
     case OP_STMT:
-      for (int i = 0; i < getChildNum(); i ++) childBits.push_back(usedBit);
+      for (size_t i = 0; i < getChildNum(); i ++) childBits.push_back(usedBit);
       break;
     case OP_READ_MEM:
       childBits.push_back(Child(0, width));
@@ -126,7 +126,7 @@ void ENode::passWidthToChild() {
     case OP_PRINTF:
     case OP_ASSERT:
     case OP_EXT_FUNC:
-      for (int i = 0; i < getChildNum(); i ++) {
+      for (size_t i = 0; i < getChildNum(); i ++) {
         childBits.push_back(getChild(i)->width);
       }
       break;
